Harden compiler detection and CFLAGS setting
Instead of guessing which compiler will be in use, we now hook in to the
distutils build process so we can choose all our extra compiler flags
once the compiler has actually been chosen.

Previously, the test on Windows would not activate if QuTiP was being
built within an MSYS bash instance, even if the MSVC toolchain was being
used, not mingw64.