Use cythonize not Cython.Distutil.build_ext
Cythonize handles dependencies between .pxd and .pyx files correctly
when cythonising multiple extension modules, whereas the (old)
Cython.Distutil.build_ext does not.  For example, if changes are made to
the Accumulator functions in `qutip/core/data/csr.pxd`, cythonize will
recognise that everything that cimports csr will need to be updated,
whereas the build_ext overrides will only re-cythonise csr.pyx.

As of Cython 0.29 there is a `new_build_ext` in the Cython source which
internally calls cythonize, but it's shadowed in the actual release.