Unify functions that return a copy of an object
This currently affects the Pauli-matrix creation functions, but is being
added to make it simpler to handle various gates being imported in from
qutip.qip as it is moved into a separate package.

The new form adds the `dtype` argument to all the function calls, backed
by a general cache.  This brings it up to feature parity with other
operator-creation functions, and the backing cache ensures we have the
same speed each time.