Import gates into core qutip
Since qutip.qip is being split out, and the gates are convenient
functions to have in general, we are copying them across.  We are not
keeping the full definitions, where each one has keyword arguments to
expand it into a given subspace, but we are bringing in expand_operator
so users can do it manually.

The general intention is that these will be good for quick-and-dirty
operations, but somebody who wants to have the full qutip.qip processing
power will install the external package.